chore: updated mods to latest versions.
We may remove the Drawers mod in an upcomming version.

Mods that conflicted with the update:
	ethereal

Mods updated this release:

	3d_armor
	airutils
	animalia
	anvil
	areas
	automobiles_pck
	bakedclay
	basic_signs
	blockexchange
	bonemeal
	carpets
	creatura
	currency
	digilines
	dmobs
	draconis
	etherium_stuff
	fakelib
	farming
	hangglider
	hbhunger
	hidroplane
	homedecor_modpack
	hopper
	hudbars
	illumination
	itemframes
	ju52
	mesecons
	mob_horse
	mobkit
	mobs
	mobs_animal
	mobs_monster
	mobs_npc
	moreblocks
	moremesecons
	moreores
	motorboat
	mtimer
	mtzip
	mydoors
	nautilus
	nether
	other_worlds
	pipeworks
	promise
	protector
	respawn
	signs_lib
	skinsdb
	steampunk_blimp
	supercub
	technic
	tpr
	trike
	unifieddyes
	unified_inventory
	worldedit
	x_bows
	xcompat

+++ mods/worldedit/.util/run_tests.sh
@@ -2,6 +2,10 @@
 tempdir=$(mktemp -d)
 confpath=$tempdir/minetest.conf
 worldpath=$tempdir/world
+modlist=(
+	worldedit
+	worldedit_commands
+)
 trap 'rm -rf "$tempdir"' EXIT
 
 [ -f worldedit/mod.conf ] || { echo "Must be run in modpack root folder." >&2; exit 1; }
@@ -11,9 +15,9 @@ if [ "$1" == "--docker" ]; then
 	command -v docker >/dev/null || { echo "Docker is not installed." >&2; exit 1; }
 	[ -d minetest_game ] || echo "A source checkout of minetest_game was not found. This can fail if your docker image does not ship a game." >&2;
 else
-	mtserver=$(command -v minetestserver)
-	[[ -z "$mtserver" && -x ../../bin/minetestserver ]] && mtserver=../../bin/minetestserver
-	[ -z "$mtserver" ] && { echo "To run the test outside of Docker, an installation of minetestserver is required." >&2; exit 1; }
+	mtserver=$(command -v luantiserver)
+	[[ -z "$mtserver" && -x ../../bin/luantiserver ]] && mtserver=../../bin/luantiserver
+	[ -z "$mtserver" ] && { echo "To run the test outside of Docker, an installation of luantiserver is required." >&2; exit 1; }
 fi
 
 mkdir $worldpath
@@ -22,19 +26,23 @@ printf '%s\n' worldedit_run_tests=true max_forceloaded_blocks=9999 >$confpath
 
 if [ -z "$mtserver" ]; then
 	chmod -R 777 $tempdir
-	[ -z "$DOCKER_IMAGE" ] && DOCKER_IMAGE="ghcr.io/minetest/minetest:master"
+	[ -n "$DOCKER_IMAGE" ] || { echo "Missing DOCKER_IMAGE env variable" >&2; exit 1; }
 	vol=(
 		-v "$confpath":/etc/minetest/minetest.conf
 		-v "$tempdir":/var/lib/minetest/.minetest
-		-v "$PWD/worldedit":/var/lib/minetest/.minetest/world/worldmods/worldedit
 	)
+	for mod in "${modlist[@]}"; do
+		vol+=(-v "$PWD/$mod":/var/lib/minetest/.minetest/world/worldmods/$mod)
+	done
 	[ -d minetest_game ] && vol+=(
 		-v "$PWD/minetest_game":/var/lib/minetest/.minetest/games/minetest_game
 	)
 	docker run --rm -i "${vol[@]}" "$DOCKER_IMAGE"
 else
 	mkdir $worldpath/worldmods
-	ln -s "$PWD/worldedit" $worldpath/worldmods/worldedit
+	for mod in "${modlist[@]}"; do
+		ln -s "$PWD/$mod" $worldpath/worldmods/$mod
+	done
 	$mtserver --config "$confpath" --world "$worldpath" --logfile /dev/null
 fi
